Add files via upload
Kickstarter is main sources of initial capital for small businesses and start-up companies that are looking to launch their first products. This project will help them understand trending projects in market with their success rates.
This project will help the management team predict the success or failure of the project before its released. It will also help them understand about the creators and backers assigned with various types of projects. This project will help the management team to make critical decisions which will support in the growth of their organization.